split navbar and header
